fix(feed): route cached plain_text through the clean HTML renderer
The detail and dashboard views render descriptions through
CleanDecorator — no [N] link markers, no footnote dump, tables
flattened. But the cached plain_text, which feeds pipe-to payloads
and the content-length filter, was still going through the stock
html2text decorator. So what you piped to your script was not what
you were reading on screen. This is not great.

Move CleanDecorator and render_clean_html from ui/utils.rs into
feed.rs and build plain_text with them at parse time. They live in
feed.rs not because they're prettier there, but because the parse
path needs them and ui already depends on feed — the reverse
dependency would have been a layering violation waiting to breed.

One text pipeline. What you see, search, filter, and pipe is now
the same string.
