+++ src/ui/dashboard.rs
@@ -491,8 +491,7 @@ fn render_preview_pane<B: Backend>(
 
     // Content
     if let Some(desc) = &item.description {
-        let raw_text =
-            crate::ui::utils::render_clean_html(desc, area.width.saturating_sub(10) as usize);
+        let raw_text = crate::feed::render_clean_html(desc, area.width.saturating_sub(10) as usize);
         let formatted = format_content_for_reading(&raw_text);
         for line in formatted.lines() {
             lines.push(Line::from(vec![Span::styled(
